Move highlight_active_function tests to call_tree
module`
